pyproject.toml
[project]
name = "strands-agentcore-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=0.1.7",
    "bedrock-agentcore-starter-toolkit>=0.1.19",
    "fastapi>=0.118.0",
    "nest-asyncio>=1.6.0",
    "ngrok>=1.5.1",
    "playwright>=1.55.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "strands-agents>=1.10.0",
    "strands-agents-tools>=0.2.9",
    "uvicorn>=0.37.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
exclude = ["uploads*", "documents*", "agent_memory*"]
